Check if internal edges are present before deleting/ removing them in extrude.

+++ client/js/editor/object.js
@@ -617,7 +617,10 @@ shapy.editor.Object.prototype.extrude = function(faces) {
       return 'internal';
     }
   });
-  goog.array.removeDuplicates(sortedEdges.internal);
+
+  if (sortedEdges.internal) {
+    goog.array.removeDuplicates(sortedEdges.internal);
+  }
 
   // Compute the vertices
   var verts = goog.array.flatten(goog.array.map(edges, function(e) {
@@ -752,9 +755,11 @@ shapy.editor.Object.prototype.extrude = function(faces) {
   }, this);
 
   // Delete internal edges
-  goog.array.forEach(sortedEdges.internal, function(e) {
-    e.delete();
-  });
+  if (sortedEdges.internal) {
+    goog.array.forEach(sortedEdges.internal, function(e) {
+      e.delete();
+    });
+  }
 
   this.dirtyMesh = true;
